Formant measurements from 564 vowel midpoints

A dataset containing formant measurements from the midpoints of 564 vowel tokens. These came from generated nonce words. This dataset is relatively clean and can be used for vowel formant example data.

midpoints

Format

A data frame with 564 rows and 12 variables:

vowel_id

a unique identifier for each vowel token

start

the start time for that vowel

end

the end time for that vowel

t

the time where formants were extracted

F1

the F1 measurement

F2

the F2 measurement

F3

the F3 measurement

F4

the F4 measurement

word

the generated nonce word I read

pre

the consonant(s) before the vowel (if any)

vowel

the vowel class, in Wells' Lexical Sets

fol

the consonant(s) after the vowel

Details

This dataset is a subset of the coronals dataframe. It only includes data from vowel midpoints. The diphthongs PRICE, MOUTH, and CHOICE are excluded. See ?coronals for more information on how the dataset was prepared.